healthcare_rag/app/core/config.py
from pydantic_settings import BaseSettings
from dotenv import load_dotenv

load_dotenv(".env")


class Settings(BaseSettings):
    POSTGRES_URL: str
    REDIS_URL: str | None = None
    ENVIRONMENT: str = "development"
    OPENAI_API_KEY: str | None = None

    class Config:
        env_file = ".env"
        env_file_encoding = "utf-8"


def get_settings() -> Settings:
    return Settings()
